requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
python-multipart==0.0.12
supabase==2.10.0
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==1.0.2
pydantic==2.9.2